StreamDataTradingSystem/aggregate2/requirements.txt
psycopg2-binary
websockets
pandas
